Move theming logic out of SEO component
This PR fixes #4 by moving the theming logic out of the badly-named `SEO` component (will fix this in a different/later PR) and into `src/html.js`, which can be roughly thought of as Gatsby's `index.html`. The theming logic will now run in the `head` section of the client-side HTML page, but not inside of the `SEO` component's `Helmet` tag, which is an approach to stuffing `head` from inside the React application tree. As mentioned in #4, we want the theming code to run _before_ the React application loads, as having it run _inside_ the React application can cause flickering.

This also fixes the broken build issue that currently exists on master - the theming logic is checking for `window`, which will fail whenever Gatsby tries to kick off an SSR-built version of the application. This is a pretty common React SSR issue, but in this case we don't even actually want the code to run during the SSR build, so we're okay to move here. Happy to explain this more to folks who are interested, but generally React + SSR stuff is weird and tricky, so I'll leave it at that for this commit :)