Remove link features
This div is currently hidden by CSS, but I've removed it in this commit because it currently embeds an `<a>` tag inside of _another_ `<a>` tag in the `Project` component, which is obviously not browser-compliant.  This is a separate commit in case we want to revert this or do a different change to handle this particular console warning.