Use collections to render lists of projects
This PR adds support for collections, which are groups of projects that are rendered in the UI. A collection can be tied to a single feature, and will automatically render the list of projects for that feature, or it can be manually added to, allowing for custom collections based on a theme.

It also adds "Layout", a data model that can be used to order collections. The "homepage" layout is a special object (that exists in staging and will be ported to production) to allow the list of collections on the root path to be dynamically updated and re-ordered.

Finally, the "Related Projects" section of the project page, which has been hardcoded with example projects, now looks for the first collection it can find that the current project exists in, and renders the remainder of the projects from that collection.

Closes #17