Fix clap component flicker issue
This commit fixes the clap component flicker issue by re-engineering the structure of the Clap component. The short version: it now uses a nifty `useSSR` hook to conditionally check if the component is being rendered in a browser. If it _is_, it renders our previous Clap component. If the component is being server-side rendered, it now falls back to rendering _just_ the Clap button (without any data or an `onClick` function). A CSS class `.Clap-is-loading` has been added as well, which hides the contents of `.Clap` and disables pointer events, while the component is loading.