Update Sites Instructions to API Token
API Tokens via the profile page are the recommended auth path by the wrangler config cli and good security hygiene. It can be confusing when a new user is looking to try Sites.

Update language and instructions to direct them to create a limited token with just Workers permissions via the Workers template.

If Global API Key is still desired for the quickstart the link to the profile should be updated at least so that it points to the API section of the updated profile section.