Remove Sites page and update nav
This commit removes the /sites page from the Workers landing page. In #151 we began redirecting /sites to Cloudflare Pages, and this commit updates the nav (header and footer) to point to Cloudflare Pages directly.

This commit also removes the CTA for people to signup for Pages. The component is still available in case we use it for a later purpose, but is no longer rendered on the site.